Split NDJSON distinct filters
